Add utility function for generating benchmarks
To be able to test potential improvements to the solver, we need to be
able to try it on a lot of different positions. Doing this by hand gets
tedious quite quickly, so we generate positions which satisfy certain
conditions: min number of moves, minimal depth needed to solve it, and
maximal depth needed to solve it.