Add transposition table.
Currently it is only used for better move ordering, because there seem
to be some issues with using the stored score for pruning. Even so, it
brings a huge cut in number of nodes visited. Additionally, we now get
a pv output again, by looking up best moves in the table.

This also changes the way "mate" scores are stored, by using the number
of generic moves, instead of the number of turns.